Reimplement GeoJSON importer as superclass of OGR
It turns out it is useful to have a string-based GeoJSON importer so the host app can import parsed JSON as opposed to only file-based JSON.
